% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_apply.R
\name{.apply_numeric_pattern}
\alias{.apply_numeric_pattern}
\title{Apply Numeric Pattern to Values}
\usage{
.apply_numeric_pattern(x, format, keep_na = FALSE)
}
\description{
Apply Numeric Pattern to Values
}
\keyword{internal}
