% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_parse.R
\name{.block_to_format}
\alias{.block_to_format}
\title{Convert a parsed block to a ks_format or ks_invalue object}
\usage{
.block_to_format(block)
}
\description{
Convert a parsed block to a ks_format or ks_invalue object
}
\keyword{internal}
