% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_parse.R
\name{.block_to_ks_datetime_format}
\alias{.block_to_ks_datetime_format}
\title{Convert VALUE block with date/time/datetime type to ks_format}
\usage{
.block_to_ks_datetime_format(block)
}
\description{
Convert VALUE block with date/time/datetime type to ks_format
}
\keyword{internal}
