% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_parse.R
\name{.block_to_ks_invalue}
\alias{.block_to_ks_invalue}
\title{Convert INVALUE block to ks_invalue}
\usage{
.block_to_ks_invalue(block)
}
\description{
Convert INVALUE block to ks_invalue
}
\keyword{internal}
