% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_parse.R
\name{.cntlout_to_ks_format}
\alias{.cntlout_to_ks_format}
\title{Convert CNTLOUT rows to ks_format}
\usage{
.cntlout_to_ks_format(rows, fmt_name, fmt_type, skipped)
}
\description{
Convert CNTLOUT rows to ks_format
}
\keyword{internal}
