% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_parse.R
\name{.datetime_format_to_text}
\alias{.datetime_format_to_text}
\title{Convert datetime ks_format to SAS-like text}
\usage{
.datetime_format_to_text(fmt, name)
}
\description{
Convert datetime ks_format to SAS-like text
}
\keyword{internal}
