% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.format_library}
\alias{.format_library}
\title{Format Library (Global Environment for Storing Formats)}
\usage{
.format_library
}
\description{
Format Library (Global Environment for Storing Formats)
}
\keyword{internal}
