% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_parse.R
\name{.format_range_bound}
\alias{.format_range_bound}
\title{Format a numeric range bound for text output}
\usage{
.format_range_bound(val, is_low = TRUE)
}
\description{
Format a numeric range bound for text output
}
\keyword{internal}
