% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_datetime.R
\name{.is_sas_datetime_format}
\alias{.is_sas_datetime_format}
\title{Check if a name corresponds to a built-in SAS datetime format}
\usage{
.is_sas_datetime_format(name)
}
\arguments{
\item{name}{Character. Format name to check.}
}
\value{
Logical.
}
\description{
Check if a name corresponds to a built-in SAS datetime format
}
\keyword{internal}
