% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_datetime.R
\name{.sas_format_defaults}
\alias{.sas_format_defaults}
\title{Default widths for SAS format names without explicit width}
\usage{
.sas_format_defaults
}
\description{
Default widths for SAS format names without explicit width
}
\keyword{internal}
