% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_datetime.R
\name{.to_r_datetime}
\alias{.to_r_datetime}
\title{Convert input to R POSIXct}
\usage{
.to_r_datetime(x, origin = "1970-01-01")
}
\arguments{
\item{x}{Input vector}

\item{origin}{Character. Origin date for numeric conversion.
Default \code{"1970-01-01"} (R epoch).}
}
\value{
POSIXct vector
}
\description{
Handles POSIXct, Date, numeric (using specified epoch origin),
and character inputs.
}
\keyword{internal}
