% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_list}
\alias{format_list}
\title{List Available Formats in Library}
\usage{
format_list()
}
\value{
Character vector of format names
}
\description{
Returns names of all formats stored in the global format library.
}
\examples{
format_list()
}
