% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_create.R
\name{print.ks_format}
\alias{print.ks_format}
\title{Print Format Object}
\usage{
\method{print}{ks_format}(x, ...)
}
\arguments{
\item{x}{A ks_format object}

\item{...}{Additional arguments (unused)}
}
\value{
The input \code{x}, returned invisibly.
}
\description{
Print Format Object
}
