% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_invalue.R
\name{print.ks_invalue}
\alias{print.ks_invalue}
\title{Print Invalue Object}
\usage{
\method{print}{ks_invalue}(x, ...)
}
\arguments{
\item{x}{A ks_invalue object}

\item{...}{Additional arguments (unused)}
}
\value{
The input \code{x}, returned invisibly.
}
\description{
Print Invalue Object
}
