% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_datetime.R
\docType{data}
\name{.sas_format_defaults}
\alias{.sas_format_defaults}
\title{Default widths for SAS format names without explicit width}
\format{
An object of class \code{list} of length 15.
}
\usage{
.sas_format_defaults
}
\description{
Default widths for SAS format names without explicit width
}
\keyword{internal}
