% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_datetime.R
\name{.to_r_date}
\alias{.to_r_date}
\title{Convert input to R Date}
\usage{
.to_r_date(x)
}
\arguments{
\item{x}{Input vector}
}
\value{
Date vector
}
\description{
Handles Date, POSIXct, numeric (SAS epoch: days since 1960-01-01),
and character inputs.
}
\keyword{internal}
