% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_datetime.R
\name{.to_r_datetime}
\alias{.to_r_datetime}
\title{Convert input to R POSIXct}
\usage{
.to_r_datetime(x)
}
\arguments{
\item{x}{Input vector}
}
\value{
POSIXct vector
}
\description{
Handles POSIXct, Date, numeric (SAS epoch: seconds since 1960-01-01 00:00:00),
and character inputs.
}
\keyword{internal}
