% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_get}
\alias{format_get}
\title{Retrieve Format from Library}
\usage{
format_get(name)
}
\arguments{
\item{name}{Character. Name of the format to retrieve}
}
\value{
A ks_format or ks_invalue object
}
\description{
Gets a format definition from the global format library.
}
\examples{
\dontrun{
sex_fmt <- format_get("sex")
}
}
