% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_apply.R
\name{fput_all}
\alias{fput_all}
\title{Apply Format and Return All Matches (Multilabel)}
\usage{
fput_all(x, format, ..., keep_na = FALSE)
}
\arguments{
\item{x}{Vector of values to format}

\item{format}{A \code{ks_format} object or a character string naming a format
in the global format library.}

\item{...}{Additional arguments for expression labels (mapped to \code{.x1},
\code{.x2}, etc.).}

\item{keep_na}{Logical. If TRUE, preserve NA in output.}
}
\value{
A list of character vectors. Each element contains all matching labels
  for the corresponding input value.
}
\description{
For multilabel formats, returns all matching labels for each input value.
Regular \code{\link{fput}} returns only the first match; this function
returns all matches as a list of character vectors.
}
\examples{
age_ml <- fnew(
  "0,5,TRUE,TRUE" = "Infant",
  "6,11,TRUE,TRUE" = "Child",
  "12,17,TRUE,TRUE" = "Teen",
  "0,17,TRUE,TRUE" = "Minor",
  "18,64,TRUE,TRUE" = "Adult",
  "65,Inf,TRUE,TRUE" = "Senior",
  name = "age_ml", type = "numeric", multilabel = TRUE
)

fput_all(c(3, 15, 25), age_ml)
# [[1]] "Infant" "Minor"
# [[2]] "Teen" "Minor"
# [[3]] "Adult"
fclear()
}
